tests/test_main_page_question.py
import pytest
from selenium import webdriver
from conftest import driver
from locators.locators_main_page_question import MainPageLocators
from pages.base_page import BasePage
from pages.main_page_question import MainPageQuestion
import time
from pages.base_page import BasePage
from urls import UrelsSamokati


class TestQuestionText:

    @pytest.mark.parametrize('num, expected_words', [
            (0, "Сутки — 400 рублей. Оплата курьеру — наличными или картой."),
            (1, "Пока что у нас так: один заказ — один самокат. Если хотите покататься с друзьями, можете просто сделать несколько заказов — один за другим."),
            (2, "Допустим, вы оформляете заказ на 8 мая. Мы привозим самокат 8 мая в течение дня. Отсчёт времени аренды начинается с момента, когда вы оплатите заказ курьеру. Если мы привезли самокат 8 мая в 20:30, суточная аренда закончится 9 мая в 20:30."),
            (3, "Только начиная с завтрашнего дня. Но скоро станем расторопнее."),
            (4, "Пока что нет! Но если что-то срочное — всегда можно позвонить в поддержку по красивому номеру 1010."),
            (5, "Самокат приезжает к вам с полной зарядкой. Этого хватает на восемь суток — даже если будете кататься без передышек и во сне. Зарядка не понадобится."),
            (6, "Да, пока самокат не привезли. Штрафа не будет, объяснительной записки тоже не попросим. Все же свои."),
            (7, "Да, обязательно. Всем самокатов! И Москве, и Московской области.")])

    def test_text_in_object_1(self, driver, num, expected_words):
        main_page = MainPageQuestion(driver)
        main_page.enter_the_page(UrelsSamokati.BASE_URL)
        assert  main_page.get_text_assert(num) == expected_words